# Remove uneeded `ORDER BY` query