--- a/app/controllers/follower_accounts_controller.rb
+++ b/app/controllers/follower_accounts_controller.rb
@@ -4,12 +4,6 @@ class FollowerAccountsController < ApplicationController
   include AccountControllerConcern
 
   def index
-    @accounts = ordered_accounts.page(params[:page]).per(FOLLOW_PER_PAGE)
-  end
-
-  private
-
-  def ordered_accounts
-    @account.followers.order('follows.created_at desc')
+    @accounts = @account.followers.page(params[:page]).per(FOLLOW_PER_PAGE)
   end
 end