--- a/app/controllers/following_accounts_controller.rb
+++ b/app/controllers/following_accounts_controller.rb
@@ -4,12 +4,6 @@ class FollowingAccountsController < ApplicationController
   include AccountControllerConcern
 
   def index
-    @accounts = ordered_accounts.page(params[:page]).per(FOLLOW_PER_PAGE)
-  end
-
-  private
-
-  def ordered_accounts
-    @account.following.order('follows.created_at desc')
+    @accounts = @account.following.page(params[:page]).per(FOLLOW_PER_PAGE)
   end
 end